backend/requirements.txt
flask>=1.1.1
mysql-connector-python>=8.0.18 
